Move links component into Navbar component
